WP_article_collector/requirements.txt
# Async support
nest_asyncio>=1.5.6
aiohttp>=3.8.3

# Data processing
pandas>=1.5.0
bs4

# Database
psycopg2-binary>=2.9.5
python-dotenv>=0.21.0

# Wikipedia APIs
requests>=2.28.1

